# src/api/openapi.json
openapi: 3.0.0
info:
  title: Calendar Booking API
  description: |-
    # Calendar Booking API v2.0.0

    Сервис бронирования времени для звонков. Слоты длительностью 15 или 30 минут 
    в зависимости от выбранного типа события.
  version: 0.0.0
tags:
  - name: Public
  - name: Owner
paths:
  /api/bookings:
    post:
      operationId: PublicAPI_createBooking
      description: |-
        Создать бронирование

        **Валидация:**
        - startAt должен соответствовать сетке слотов (кратен durationMinutes)
        - Время должно быть в рамках OwnerSettings.workDayStart - workDayEnd
        - Не должно быть пересечений с существующими бронированиями (409 Conflict)

        **Правило занятости:**
        На одно и то же время нельзя создать два бронирования — даже если они
        относятся к разным типам событий.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: 'Ошибка: недопустимое время слота (не кратно durationMinutes или вне рабочих часов)'
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/InvalidSlotTimeError'
                  - $ref: '#/components/schemas/ValidationError'
        '404':
          description: 'Ошибка: ресурс не найден'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '409':
          description: 'Ошибка: слот уже занят (пересечение с существующим бронированием)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
      tags:
        - Public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
  /api/event-types:
    get:
      operationId: PublicAPI_listEventTypes
      description: Получить список доступных типов событий
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
      tags:
        - Public
  /api/owner/bookings:
    get:
      operationId: OwnerAPI_listBookings
      description: |-
        Получить список бронирований (с фильтрами).

        Без фильтров: возвращаются бронирования за сегодня и будущие (startAt >= startOfToday),
        отсортированные по startAt ASC.

        Возвращает все бронирования, включая orphan (с удалёнными типами событий).
      parameters:
        - $ref: '#/components/parameters/ListBookingsRequest.eventTypeId'
        - $ref: '#/components/parameters/ListBookingsRequest.startDate'
        - $ref: '#/components/parameters/ListBookingsRequest.endDate'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
      tags:
        - Owner
  /api/owner/bookings/{id}:
    delete:
      operationId: OwnerAPI_cancelBooking
      description: Отменить бронирование. Возвращает 204 No Content при успехе.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: 'Ошибка: ресурс не найден'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
      tags:
        - Owner
  /api/owner/event-types:
    post:
      operationId: OwnerAPI_createEventType
      description: |-
        Создать тип события

        **Ограничение:** нельзя создать более одного типа с duration=15
        и более одного типа с duration=30.

        **Сид данных:** при первом запуске автоматически создаются:
        - "Встреча 15 минут" (15 минут)
        - "Встреча 30 минут" (30 минут)
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: 'Ошибка: тип события с такой длительностью уже существует'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicateDurationError'
      tags:
        - Owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTypeRequest'
  /api/owner/event-types/{id}:
    patch:
      operationId: OwnerAPI_updateEventType
      description: |-
        Обновить тип события (частичное обновление).
        Длительность изменить нельзя после создания.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: 'Ошибка: ресурс не найден'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
      tags:
        - Owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventTypeRequest'
    delete:
      operationId: OwnerAPI_deleteEventType
      description: |-
        Удалить тип события. Возвращает 204 No Content при успехе.

        **Примечание:** существующие бронирования этого типа остаются в системе
        (orphan bookings), но новые бронирования этого типа станут невозможны.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: 'Ошибка: ресурс не найден'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
      tags:
        - Owner
  /api/owner/settings:
    patch:
      operationId: OwnerAPI_updateSettings
      description: Обновить настройки владельца (частичное обновление)
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerSettings'
      tags:
        - Owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOwnerSettingsRequest'
  /api/settings:
    get:
      operationId: PublicAPI_getSettings
      description: Получить настройки владельца (профиль, аватар, часы работы)
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerSettings'
      tags:
        - Public
  /api/slots:
    get:
      operationId: PublicAPI_listSlots
      description: |-
        Получить слоты за период для выбранного типа события

        Возвращает только будущие слоты (startAt > now()).

        Используется дважды:
        - С диапазоном на весь видимый месяц (например, 2026-03-01..2026-03-31) — для отображения 
        счётчика свободных слотов под каждым днём календаря
        - С диапазоном на один день (например, 2026-03-28..2026-03-28) — для отображения 
        списка слотов справа
      parameters:
        - $ref: '#/components/parameters/ListSlotsRequest.eventTypeId'
        - $ref: '#/components/parameters/ListSlotsRequest.startDate'
        - $ref: '#/components/parameters/ListSlotsRequest.endDate'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
      tags:
        - Public
components:
  parameters:
    ListBookingsRequest.endDate:
      name: endDate
      in: query
      required: false
      description: 'Фильтр: бронирования до этой даты включительно. Формат: YYYY-MM-DD'
      schema:
        type: string
        format: date
      explode: false
    ListBookingsRequest.eventTypeId:
      name: eventTypeId
      in: query
      required: false
      description: Фильтр по типу события
      schema:
        type: string
      explode: false
    ListBookingsRequest.startDate:
      name: startDate
      in: query
      required: false
      description: 'Фильтр: бронирования с этой даты. Формат: YYYY-MM-DD'
      schema:
        type: string
        format: date
      explode: false
    ListSlotsRequest.endDate:
      name: endDate
      in: query
      required: true
      description: 'Конец периода (дата, включительно). Формат: YYYY-MM-DD, например 2026-03-31'
      schema:
        type: string
        format: date
      explode: false
    ListSlotsRequest.eventTypeId:
      name: eventTypeId
      in: query
      required: true
      description: ID типа события (определяет шаг сетки слотов)
      schema:
        type: string
      explode: false
    ListSlotsRequest.startDate:
      name: startDate
      in: query
      required: true
      description: 'Начало периода. Формат: YYYY-MM-DD, например 2026-03-01'
      schema:
        type: string
        format: date
      explode: false
  schemas:
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - eventTypeName
        - guestName
        - guestEmail
        - startAt
        - endAt
        - createdAt
      properties:
        id:
          type: string
        eventTypeId:
          type: string
          description: |-
            ID типа события. 
            Может ссылаться на удалённый тип (orphan booking после DELETE /api/event-types/{id}).
        eventTypeName:
          type: string
          description: Название типа события. Сохраняется на момент создания (snapshot).
        guestName:
          type: string
          description: Имя гостя
        guestEmail:
          type: string
          description: Email гостя
        startAt:
          type: string
          format: date-time
          description: Начало бронирования
        endAt:
          type: string
          format: date-time
          description: Конец бронирования
        createdAt:
          type: string
          format: date-time
          description: Дата создания бронирования
      description: Бронирование — запись гостя на выбранный слот
    ConflictError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - SLOT_ALREADY_BOOKED
        message:
          type: string
          enum:
            - Выбранный слот уже занят. Пожалуйста, выберите другое время.
      description: 'Ошибка: слот уже занят (пересечение с существующим бронированием)'
    CreateBookingRequest:
      type: object
      required:
        - eventTypeId
        - guestName
        - guestEmail
        - startAt
      properties:
        eventTypeId:
          type: string
          description: ID выбранного типа события
        guestName:
          type: string
          description: Имя гостя
        guestEmail:
          type: string
          description: Email гостя
        startAt:
          type: string
          format: date-time
          description: Начало выбранного слота (строго кратно durationMinutes типа)
      description: |-
        Запрос на создание бронирования

        **Валидация startAt:**
        - Для типа 15 минут: время должно быть кратно 15 минутам (XX:00, XX:15, XX:30, XX:45)
        - Для типа 30 минут: время должно быть кратно 30 минутам (XX:00, XX:30)
        - Время должно быть в рамках OwnerSettings.workDayStart - workDayEnd
    CreateEventTypeRequest:
      type: object
      required:
        - name
        - description
        - durationMinutes
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: number
          enum:
            - 15
            - 30
          description: Длительность — только 15 или 30 минут
      description: |-
        Запрос на создание типа события

        **Ограничение:** нельзя создать более одного типа с duration=15
        и более одного типа с duration=30.
    DuplicateDurationError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - DUPLICATE_DURATION
        message:
          type: string
          enum:
            - Тип события с такой длительностью уже существует. Разрешено только по одному типу для 15 и 30 минут.
      description: 'Ошибка: тип события с такой длительностью уже существует'
    EventType:
      type: object
      required:
        - id
        - name
        - description
        - durationMinutes
      properties:
        id:
          type: string
        name:
          type: string
          description: Название типа события (например, "Встреча 15 минут", "Встреча 30 минут")
        description:
          type: string
          description: Описание типа события
        durationMinutes:
          type: number
          enum:
            - 15
            - 30
          description: Длительность в минутах — только 15 или 30
      description: Тип события — определяет длительность встречи (15 или 30 минут)
    InvalidSlotTimeError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - INVALID_SLOT_TIME
        message:
          type: string
          enum:
            - Недопустимое время начала слота. Время должно соответствовать сетке и быть в рамках рабочих часов.
      description: 'Ошибка: недопустимое время слота (не кратно durationMinutes или вне рабочих часов)'
    NotFoundError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - NOT_FOUND
        message:
          type: string
      description: 'Ошибка: ресурс не найден'
    OwnerSettings:
      type: object
      required:
        - name
        - workDayStart
        - workDayEnd
      properties:
        name:
          type: string
          description: Имя владельца (отображается гостям)
        avatarUrl:
          type: string
          description: URL аватара владельца
        workDayStart:
          type: string
          description: 'Начало рабочего дня. Формат: HH:mm, 24-часовой, например "09:00"'
        workDayEnd:
          type: string
          description: 'Конец рабочего дня. Формат: HH:mm, 24-часовой, например "18:00"'
      description: |-
        Настройки владельца календаря (профиль и часы работы).
        Владелец один и предзадан. В API возвращается без id, маршрут всегда /api/settings.
        Дефолтные значения: workDayStart="09:00", workDayEnd="18:00".
    Slot:
      type: object
      required:
        - startAt
        - endAt
        - isBooked
      properties:
        startAt:
          type: string
          format: date-time
          description: Начало слота
        endAt:
          type: string
          format: date-time
          description: Конец слота (startAt + EventType.durationMinutes)
        isBooked:
          type: boolean
          description: |-
            Занят ли слот любым бронированием.
            Только факт занятости. В PublicAPI данные о госте не возвращаются никогда.
      description: Слот для бронирования — временное окно для конкретного типа события
    UpdateEventTypeRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      description: |-
        Запрос на обновление типа события (частичное обновление).
        Длительность изменить нельзя — только name и description.
    UpdateOwnerSettingsRequest:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
        workDayStart:
          type: string
        workDayEnd:
          type: string
      description: Запрос на обновление настроек владельца
    ValidationError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - VALIDATION_ERROR
        message:
          type: string
        details:
          type: array
          items:
            type: string
      description: Ошибка валидации